+++ NEWS.md
@@ -3,6 +3,7 @@
 * Use GPL (>= 3) License.
 * drop `terra_` prefix and move `rast_` functions into `scratch` where we test the LOO method. * Add `sf_` prefix to vector based approaches.
 * Add datasets and dataset loading function
+* Add legend to habitat buffer barrier plot - [#66](https://github.com/urbio-ecology/urbioconnect/issues/66)
 
 # urbioconnect 0.1.0